Refactor: Extract CaseRepository from CaseController
This commit extracts the case data access logic from the `CaseController` into a new `CCDCaseRepository` class.

This change improves the separation of concerns by making the `CaseController` responsible only for handling HTTP requests and leaving the data access details to the repository. This makes the code cleaner and easier to maintain.

- Moved case retrieval and serialization logic to `CCDCaseRepository`.

- Injected `CCDCaseRepository` into `CaseController` and `CaseEventHistoryService`.

- Updated method signatures to reflect the new design.
